[refactor] configuration setup
