+++ src/tests/lib/config.test.ts
@@ -1,12 +1,12 @@
 import test from 'ava';
 import * as path from 'path';
 
+import Config from '../../config';
 import { ETHEREUM_DATADIR } from '../../data';
-import { ConfigParser } from '../../lib/config';
 
 test('default config should be valid', async t => {
   const defaultConfigPath = path.join(ETHEREUM_DATADIR, 'double.yaml');
-  const cfg = await ConfigParser.parseFromFile(defaultConfigPath);
+  const cfg = await Config.parseFromFile(defaultConfigPath);
 
   t.is(cfg.project, 'default');
   t.is(cfg.chain, 'ethereum');
